api/requirements.txt
beautifulsoup4~=4.13.4
requests~=2.32.4
playwright~=1.54.0
pyfiglet~=1.0.3
fastapi~=0.116.1
uvicorn[standard]~=0.30.6
passlib[bcrypt]~=1.7.4
PyJWT~=2.9.0
python-multipart
